Revert "fix: 모든 경고 제거"
This reverts commit 7ad52e61c65cca4e9e41f30a0e638baedcdf969e.

+++ Mundi/Source/Editor/Gizmo/GizmoArrowComponent.cpp
@@ -87,7 +87,7 @@ void UGizmoArrowComponent::CollectMeshBatches(TArray<FMeshBatchElement>& OutMesh
 	}
 
 	// 화면에 그려지는 크기 설정
-	SetDrawScale(static_cast<float>(View->ViewRect.Width()), static_cast<float>(View->ViewRect.Height()), View->ViewMatrix, View->ProjectionMatrix);
+	SetDrawScale(View->ViewRect.Width(), View->ViewRect.Height(), View->ViewMatrix, View->ProjectionMatrix);
 
 	// --- 사용할 머티리얼 결정 ---
 	UMaterialInterface* MaterialToUse = GizmoMaterial;
