+++ Mundi/Source/Runtime/Renderer/SceneView.cpp
@@ -32,8 +32,8 @@ FSceneView::FSceneView(FMinimalViewInfo* InMinimalViewInfo, URenderSettings* InR
 	ProjectionMatrix = FMatrix::CreateProjectionMatrix(
 		InMinimalViewInfo->FieldOfView,
 		InMinimalViewInfo->AspectRatio,
-		static_cast<float>(InMinimalViewInfo->ViewRect.Width()),
-		static_cast<float>(InMinimalViewInfo->ViewRect.Height()),
+		InMinimalViewInfo->ViewRect.Width(),
+		InMinimalViewInfo->ViewRect.Height(),
 		InMinimalViewInfo->NearClip,
 		InMinimalViewInfo->FarClip,
 		InMinimalViewInfo->ZoomFactor,
